[#245] catch all exceptions during yaml parsing
- Move showErrorMessage method from LSpUtils to
ClangdConfigurationFileManager since it's only used here. Despite that
we do not want to publish the LspUtils when making the
org.eclipse.cdt.lsp.clangd package public API.
