+++ bundles/org.eclipse.cdt.lsp.clangd/src/org/eclipse/cdt/lsp/internal/clangd/editor/LspEditorUiMessages.java
@@ -46,7 +46,7 @@ public class LspEditorUiMessages extends NLS {
 	public static String LspEditorPreferencePage_completion_default;
 	public static String LspEditorPreferencePage_select_clangd_executable;
 
-	public static String CProjectChangeMonitor_yaml_scanner_error;
-	public static String CProjectChangeMonitor_yaml_scanner_error_message;
+	public static String ClangdConfigurationFileManager_yaml_error;
+	public static String ClangdConfigurationFileManager_yaml_error_message;
 
 }
