+++ bundles/org.eclipse.cdt.lsp.clangd/src/org/eclipse/cdt/lsp/internal/clangd/editor/LspEditorUiMessages.properties
@@ -34,5 +34,5 @@ LspEditorPreferencePage_completion_bundled=Bundled
 LspEditorPreferencePage_completion_default=Default
 LspEditorPreferencePage_select_clangd_executable=Select clangd executable
 
-CProjectChangeMonitor_yaml_scanner_error=Yaml Scanner Error
-CProjectChangeMonitor_yaml_scanner_error_message=Error while reading:
+ClangdConfigurationFileManager_yaml_error=Yaml Error
+ClangdConfigurationFileManager_yaml_error_message=Error while reading:\n
